+++ backend/src/neuroagent/app/dependencies.py
@@ -1,6 +1,7 @@
 """App dependencies."""
 
 import logging
+from datetime import datetime, timezone
 from functools import cache
 from typing import Annotated, Any, AsyncIterator
 
@@ -31,7 +32,6 @@ from neuroagent.tools import (
     MEModelGetOneTool,
     MorphologyFeatureTool,
     MorphologyViewerTool,
-    NowTool,
     PlotGeneratorTool,
     ResolveEntitiesTool,
     SCSGetAllTool,
@@ -180,7 +180,7 @@ def get_tool_list() -> list[type[BaseTool]]:
         PlotGeneratorTool,
         MorphologyViewerTool,
         WebSearchTool,
-        NowTool,
+        # NowTool,
         # WeatherTool,
         # RandomPlotGeneratorTool,
     ]
@@ -210,15 +210,18 @@ def get_starting_agent(
 ) -> Agent:
     """Get the starting agent."""
     logger.info(f"Loading model {settings.openai.model}.")
-    base_instructions = """You are a helpful assistant helping scientists with neuro-scientific questions.
+    base_instructions = f"""You are a helpful assistant helping scientists with neuro-scientific questions.
+                The current date and time is {datetime.now(timezone.utc).isoformat()}.
                 You must always specify in your answers from which brain regions the information is extracted.
-                Do no blindly repeat the brain region requested by the user, use the output of the tools instead.
-                We provide a description of the platform, the open brain platform allows an atlas driven exploration of the mouse brain with different artifacts related to experimental and model data and more specifically neuron morphology
+                Do not blindly repeat the brain region requested by the user, use the output of the tools instead.
+                You are embedded in a platform called the Open Brain Platform.
+                The Open Brain Platform allows an atlas driven exploration of mouse, rat and human brain data with different artifacts related to experimental and model data, more specifically: neuron morphology
                 (neuron structure including axons, soma and dendrite), electrophysiological recording (ie the electrical behavior of the neuron), ion channel, neuron density, bouton density, synapses, connections, electrical models also referred to as e-models, me-models which is the model of neuron with a specific morphology and electrical type, and the synaptome dictating how neurons are connected together.
-                The platform also allows user to explore and build digital brain models at different scales ranging from molecular level to single neuron and larger circuits and brain regions.
-                Users can also customize the models or create their own ones and change the cellular composition, and then run simulation experiments and perform analysis.
-                The models currently available on the platform are the metabolism and NGV unit as a notebook, and the single neuron, synaptome simulation. The other models will be released later starting with microcirctuits paired neurons and them brain region, brain system and whole brain.
-                The platform has a lot many notebooks that can be downloaded and executed remotely for now. A feature to run them on the platform will be available soon"""
+                The platform also allows users to explore and build digital brain models at different scales ranging from molecular level to single neuron and larger circuits and brain regions.
+                Users can also customize the models or create their own by changing the cellular composition, to then run simulation experiments and perform analysis.
+                The models currently available on the platform are the metabolism and NGV unit as a notebook, and the single neuron, synaptome simulation. The other models will be released later starting with microcircuits paired neurons and then brain region, brain system and whole brain.
+                The platform has many notebooks that can be downloaded and executed remotely for now. A feature to run them on the platform will be available soon.
+                The platform has an AI Assistant for literature search allowing users to identify articles related to the brain area and artifacts they are interested in. At a later stage, the AI assistant will be further developed to access specific tools on the platform."""
 
     storage_instructions = (
         f"All files in storage can be viewed under {settings.misc.frontend_url}/viewer/{{storage_id}}. "