+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Return only one suggestion.
+- Update system prompt.
+- Add current time and date in system prompt.
 ### Added
 - Every tool output schema to the OpenAPI json + swagger.
 - Header containing rate limit info.